requirements.txt
discord.py
gspread
oauth2client
pytz
aiofiles
python-dotenv
gspread_asyncio
aiohttp